+++ apps/benchmarking-test-app/src/graphicsBenchmarks/meshes/cube.ts
